+++ data-raw/SAMPSON.DAT.R
@@ -1,0 +1,1 @@
+file.copy("data-raw/SAMPSON.DAT", "inst/extdata/")