+++ data-raw/SAMPSON.R
@@ -1,0 +1,1 @@
+file.copy("data-raw/SAMPSON.NET", "inst/extdata/")