+++ data-raw/SAMPSONL.R
@@ -1,0 +1,1 @@
+file.copy("data-raw/SAMPSONL.NET", "inst/extdata/")