+++ data-raw/fakesurvey.R
@@ -1,3 +1,4 @@
 rm(list=ls())
 fakesurvey <- read.csv("data-raw/fakesurvey.csv")
+fakesurvey$note <- as.character(fakesurvey$note)
 save(fakesurvey, file = "data/fakesurvey.rdata")